requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
fastapi==0.123.8
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.3.5
openai==2.9.0
packaging==25.0
postgrest==2.25.0
propcache==0.4.1
psycopg==3.3.1
psycopg-binary==3.3.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
python-telegram-bot==21.6
PyYAML==6.0.3
realtime==2.25.0
requests==2.32.5
sniffio==1.3.1
starlette==0.50.0
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
